Fully working waveform
